pyproject.toml
[project]
name = "ghidra-unstrip-venv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ghidra-stubs>=11.4",
    "lief>=0.17.0",
    "pyinstaller>=6.16.0",
]
